fix(S2-Q): LlmBlock holds Arc<dyn ProviderAdmin>; block-llm compiles on wasm32
Audit finding: "LlmBlock hard-wired to concrete ProviderLlmService, creating
the llm/block-llm feature knot and excluding the LLM block from all wasm32
targets" (medium/architecture/large).

Split the provider-management surface (configure / providers_snapshot /
discover_models) into a new `ProviderAdmin` trait. `LlmBlock` now holds
`Arc<dyn ProviderAdmin>` instead of the concrete, reqwest/tokio-backed
`ProviderLlmService`. Chat/list/status traffic already went through
`ctx.call_block("wafer-run/llm", ...)` (the MultiBackendLlmService router), so
the block never needed a concrete LlmService handle — only the admin surface.

- New `blocks/llm/provider_admin.rs`: `ProviderAdmin` trait (MaybeSend +
  MaybeSync) + `NoopProviderAdmin` (always compiled).
- `providers/` split: `providers::config` (pure ProviderConfig/Protocol)
  compiles on every target; `ProviderLlmService` + its HTTP codecs move into
  `providers/service.rs`, gated `feature = "llm"`. `impl ProviderAdmin for
  ProviderLlmService` lives there.
- `mod llm` now gated on `block-llm` (not `llm`); the `llm` feature shrinks to
  "native provider backend" (the reqwest/tokio HTTP stack only).
- Mechanical registration updates: builder.rs native path passes the concrete
  service (or NoopProviderAdmin for a native block-llm-without-llm build);
  wasm32 `register_all_static_blocks` now registers `suppers-ai/llm` with
  NoopProviderAdmin; `all_block_infos`/`register_llm` take Arc<dyn ProviderAdmin>.
- solobase-web enables `block-llm`: the chat UI + thread-persistence feature
  block now runs in the browser, backed by the BrowserLlmService already
  registered on the router.

reload_provider_service / migrate_legacy_providers now take
`&dyn ProviderAdmin`. All 120 llm-block tests pass; solobase-web compiles on
wasm32-unknown-unknown.
